fix: `events/[id]` route
